chore: rename application
