+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include kwalitee/schemas/*
+include quiver/schemas/*
