+++ README.md
@@ -1,10 +1,10 @@
-# kwalitee-dashboard-back-end
+# quiver-dashboard-back-end
 
-[![CircleCI](https://circleci.com/gh/OCR-D/kwalitee-dashboard-back-end/tree/main.svg?style=svg)](https://circleci.com/gh/OCR-D/kwalitee-dashboard-back-end/tree/main)
+[![CircleCI](https://circleci.com/gh/OCR-D/quiver-dashboard-back-end/tree/main.svg?style=svg)](https://circleci.com/gh/OCR-D/quiver-dashboard-back-end/tree/main)
 
 The back end of the OCR-D quality dashboard webapp.
 
-The webapp is available over at the [OCR-D website](https://ocr-d.de/kwalitee/).
+The webapp is available over at the [OCR-D website](https://ocr-d.de/quiver/).
 
 ## Dependencies
