+++ quiver/schemas/repos.json
@@ -2,7 +2,7 @@
     "$schema": "https://json-schema.org/draft/2020-12/schema",
     "$id": "https://ocr-d.de/repos.schema.json",
     "title": "Repository Information",
-    "description": "Information needed on projects for the OCR-D Kwalitee Dashboard",
+    "description": "Information needed on projects for the OCR-D quiver Dashboard",
     "type": "array",
     "items": {
         "type": "object",
